refactor: rename the 'application' pathname e classes to 'presentation.presenters'
