+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/CategoryService.java
@@ -1,6 +1,6 @@
 package br.com.fiap.grupo30.fastfood.domain.services;
 
-import br.com.fiap.grupo30.fastfood.application.dto.CategoryDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.CategoryDTO;
 import java.util.List;
 
 public interface CategoryService {
