+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/CustomerService.java
@@ -1,6 +1,6 @@
 package br.com.fiap.grupo30.fastfood.domain.services;
 
-import br.com.fiap.grupo30.fastfood.application.dto.CustomerDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.CustomerDTO;
 
 public interface CustomerService {
