+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/ProductService.java
@@ -1,6 +1,6 @@
 package br.com.fiap.grupo30.fastfood.domain.services;
 
-import br.com.fiap.grupo30.fastfood.application.dto.ProductDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.ProductDTO;
 import java.util.List;
 
 public interface ProductService {
