+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/impl/CustomerServiceImpl.java
@@ -1,15 +1,15 @@
 package br.com.fiap.grupo30.fastfood.domain.services.impl;
 
-import br.com.fiap.grupo30.fastfood.application.dto.CustomerDTO;
-import br.com.fiap.grupo30.fastfood.application.exceptions.InvalidCpfException;
-import br.com.fiap.grupo30.fastfood.application.exceptions.ResourceConflictException;
-import br.com.fiap.grupo30.fastfood.application.exceptions.ResourceNotFoundException;
-import br.com.fiap.grupo30.fastfood.application.mapper.impl.CustomerDTOMapper;
-import br.com.fiap.grupo30.fastfood.application.mapper.impl.CustomerMapper;
 import br.com.fiap.grupo30.fastfood.domain.services.CustomerService;
 import br.com.fiap.grupo30.fastfood.domain.vo.CPF;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CustomerEntity;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories.CustomerRepository;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.CustomerDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.exceptions.InvalidCpfException;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.exceptions.ResourceConflictException;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.exceptions.ResourceNotFoundException;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.impl.CustomerDTOMapper;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.impl.CustomerMapper;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
