+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/usecases/customer/FindCustomerByCpfUseCase.java
@@ -1,7 +1,7 @@
 package br.com.fiap.grupo30.fastfood.domain.usecases.customer;
 
-import br.com.fiap.grupo30.fastfood.application.dto.CustomerDTO;
 import br.com.fiap.grupo30.fastfood.domain.services.CustomerService;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.CustomerDTO;
 import org.springframework.stereotype.Component;
 
 @Component
