+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/usecases/order/SubmitOrderUseCase.java
@@ -1,10 +1,10 @@
 package br.com.fiap.grupo30.fastfood.domain.usecases.order;
 
-import br.com.fiap.grupo30.fastfood.application.dto.OrderDTO;
-import br.com.fiap.grupo30.fastfood.application.exceptions.ResourceNotFoundException;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.OrderEntity;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.OrderStatus;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories.OrderRepository;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.OrderDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
