+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/usecases/product/ListAllCategoriesInMenuUseCase.java
@@ -1,7 +1,7 @@
 package br.com.fiap.grupo30.fastfood.domain.usecases.product;
 
-import br.com.fiap.grupo30.fastfood.application.dto.CategoryDTO;
 import br.com.fiap.grupo30.fastfood.domain.services.CategoryService;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.CategoryDTO;
 import java.util.List;
 import org.springframework.stereotype.Component;
