+++ src/main/java/br/com/fiap/grupo30/fastfood/infrastructure/persistence/entities/ProductEntity.java
@@ -1,6 +1,6 @@
 package br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities;
 
-import br.com.fiap.grupo30.fastfood.application.dto.ProductDTO;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.ProductDTO;
 import jakarta.persistence.*;
 import java.time.Instant;
 import lombok.AllArgsConstructor;
