+++ src/main/java/br/com/fiap/grupo30/fastfood/presentation/controllers/ProductController.java
@@ -1,7 +1,7 @@
 package br.com.fiap.grupo30.fastfood.presentation.controllers;
 
-import br.com.fiap.grupo30.fastfood.application.dto.ProductDTO;
 import br.com.fiap.grupo30.fastfood.domain.usecases.product.*;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.dto.ProductDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
