+++ src/main/java/br/com/fiap/grupo30/fastfood/presentation/presenters/dto/AddOrderProductRequest.java
@@ -1,4 +1,4 @@
-package br.com.fiap.grupo30.fastfood.application.dto;
+package br.com.fiap.grupo30.fastfood.presentation.presenters.dto;
 
 import lombok.Data;
