+++ src/main/java/br/com/fiap/grupo30/fastfood/presentation/presenters/mapper/Mapper.java
@@ -1,4 +1,4 @@
-package br.com.fiap.grupo30.fastfood.application.mapper;
+package br.com.fiap.grupo30.fastfood.presentation.presenters.mapper;
 
 public interface Mapper<S, T> {
     T mapTo(S source);
