+++ src/main/java/br/com/fiap/grupo30/fastfood/presentation/presenters/mapper/impl/CategoryMapper.java
@@ -1,8 +1,8 @@
-package br.com.fiap.grupo30.fastfood.application.mapper.impl;
+package br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.impl;
 
-import br.com.fiap.grupo30.fastfood.application.mapper.Mapper;
 import br.com.fiap.grupo30.fastfood.domain.Category;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CategoryEntity;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.Mapper;
 import org.springframework.stereotype.Component;
 
 @Component
