+++ src/main/java/br/com/fiap/grupo30/fastfood/presentation/presenters/mapper/impl/ProductMapper.java
@@ -1,8 +1,8 @@
-package br.com.fiap.grupo30.fastfood.application.mapper.impl;
+package br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.impl;
 
-import br.com.fiap.grupo30.fastfood.application.mapper.Mapper;
 import br.com.fiap.grupo30.fastfood.domain.Product;
 import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.ProductEntity;
+import br.com.fiap.grupo30.fastfood.presentation.presenters.mapper.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
